Make post's date output configurable via _config.yml, closes #164
A new config variable 'date_format' is introduced in _config.yml.
It can either be set to "ordinal" to use the current format or it
can be given a string complying to strftime() format identifiers.
