+++ _config.yml
@@ -8,6 +8,11 @@ subtitle: A blogging framework for hackers.
 author: Your Name
 simple_search: http://google.com/search
 
+# Default date format is "ordinal" (resulting in "July 22nd 2007")
+# You can customize the format as defined in
+# http://www.ruby-doc.org/core-1.9.2/Time.html#method-i-strftime
+date_format: "ordinal"
+
 # RSS / Email (optional) subscription links (change if using something like Feedburner)
 subscribe_rss: /atom.xml
 subscribe_email:
